Второе задание
