--- TaskManager/TaskManager/Main.aspx
@@ -1,30 +1,0 @@
-﻿<%@ Page MasterPageFile="~/Site1.Master" Language="C#" AutoEventWireup="true" CodeBehind="Main.aspx.cs" Inherits="TaskManager.Main" %>
-
-
-
-<asp:Content ContentPlaceHolderID="Content" runat="server">
-    <form action="">
-        <div id="myDIV" class="header">
-            <h2>My To Do List</h2>
-            <asp:TextBox ID="newTask" placeHolder="Title..." runat="server"></asp:TextBox>
-            <asp:Button ID="Button1" runat="server" CssClass="addBtn" Text="Button" />
-        </div>
-    </form>
-    <ul id="myUL">
-        <asp:Repeater ID="taskRepeater" runat="server">
-            <ItemTemplate>
-                <li>
-                    <div class="task">
-                        <%# Eval("Ttile") %>
-                    </div>
-                    <div class="action">
-                        <a href="#"><i class="fa fa-edit"></i></a>
-                    </div>
-                    <div class="action">
-                        <a href="#"><i class="fa fa-trash-alt"></i></a>
-                    </div>
-                </li>
-            </ItemTemplate>
-        </asp:Repeater>
-    </ul>
-</asp:Content>
