+++ TaskManager/TaskManager/Site1.Master
@@ -1,113 +1,111 @@
 ﻿<%@ Master Language="C#" AutoEventWireup="true" CodeBehind="Site1.master.cs" Inherits="TaskManager.Site1" %>
 
+
 <!DOCTYPE html>
 <html>
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <head>
+        <meta charset="utf-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1">
 
-    <title>ToDo Manager</title>
+        <title>TaskManager</title>
 
-    <!-- Fonts -->
-    <link href="https://fonts.googleapis.com/css?family=Nunito:200,600" rel="stylesheet">
-    <script src="https://kit.fontawesome.com/8500f75e5b.js" crossorigin="anonymous"></script>
+        <!-- Fonts -->
+        <link href="https://fonts.googleapis.com/css?family=Nunito:200,600" rel="stylesheet">
+        <script src="https://kit.fontawesome.com/8500f75e5b.js" crossorigin="anonymous"></script>
 
-    <!-- Styles -->
-    <style>
-        /* Include the padding and border in an element's total width and height */
-        * {
-            box-sizing: border-box;
-            font-family: Nunito, san-serif;
-        }
-        /* Remove margins and padding from the list */
-        ul {
-            margin: 0;
-            padding: 0;
-        }
+        <!-- Styles -->
+        <style>
+            /* Include the padding and border in an element's total width and height */
+            * {
+              box-sizing: border-box;
+              font-family: Nunito, san-serif;
+            }
+            /* Remove margins and padding from the list */
+            ul {
+              margin: 0;
+              padding: 0;
+            }
             /* Style the list items */
             ul li {
-                position: relative;
-                padding: 12px 8px 12px 40px;
-                background: #eee;
-                font-size: 18px;
-                transition: 0.2s;
-                display: flex;
-                flex-flow: row nowrap;
-                align-items: center;
-                justify-content: flex-start;
-                /* make the list items unselectable */
-                -webkit-user-select: none;
-                -moz-user-select: none;
-                -ms-user-select: none;
-                user-select: none;
+              position: relative;
+              padding: 12px 8px 12px 40px;
+              background: #eee;
+              font-size: 18px;
+              transition: 0.2s;
+              display: flex;
+              flex-flow: row nowrap;
+              align-items: center;
+              justify-content: flex-start;
+              /* make the list items unselectable */
+              -webkit-user-select: none;
+              -moz-user-select: none;
+              -ms-user-select: none;
+              user-select: none;
             }
-                /* Set all odd list items to a different color (zebra-stripes) */
-                ul li:nth-child(odd) {
-                    background: #f9f9f9;
-                }
-                /* Darker background-color on hover */
-                ul li:hover {
-                    background: #ddd;
-                }
-
-                ul li .task {
-                    flex-grow: 1;
-                }
-
-                ul li .action {
-                    margin: 5px 15px;
-                    vertical-align: middle;
-                }
-
-                ul li a {
-                    color: grey;
-                }
-
-        /* Style the header */
-        .header {
-            background-color: #f44336;
-            padding: 30px 40px;
-            color: white;
-            text-align: center;
-        }
+            /* Set all odd list items to a different color (zebra-stripes) */
+            ul li:nth-child(odd) {
+              background: #f9f9f9;
+            }
+            /* Darker background-color on hover */
+            ul li:hover {
+              background: #ddd;
+            }
+            ul li .task {
+                flex-grow: 1;
+            }
+            ul li .action {
+                margin: 5px 15px;
+                vertical-align: middle;
+            }
+            ul li a {
+                color: grey;
+            }
+         
+            /* Style the header */
+            .header {
+              background-color: #f44336;
+              padding: 30px 40px;
+              color: white;
+              text-align: center;
+            }
             /* Clear floats after the header */
             .header:after {
-                content: "";
-                display: table;
-                clear: both;
+              content: "";
+              display: table;
+              clear: both;
             }
-        /* Style the input */
-        input {
-            margin: 0;
-            border: none;
-            border-radius: 0;
-            width: 75%;
-            padding: 10px;
-            float: left;
-            font-size: 16px;
-        }
-        /* Style the "Add" button */
-        .addBtn {
-            padding: 10px;
-            width: 25%;
-            background: #d9d9d9;
-            color: #555;
-            float: left;
-            text-align: center;
-            font-size: 16px;
-            cursor: pointer;
-            transition: 0.3s;
-            border-radius: 0;
-            border: none;
-        }
-
+            /* Style the input */
+            input {
+              margin: 0;
+              border: none;
+              border-radius: 0;
+              width: 75%;
+              padding: 10px;
+              float: left;
+              font-size: 16px;
+            }
+            /* Style the "Add" button */
+            .addBtn {
+              padding: 10px;
+              width: 25%;
+              background: #d9d9d9;
+              color: #555;
+              float: left;
+              text-align: center;
+              font-size: 16px;
+              cursor: pointer;
+              transition: 0.3s;
+              border-radius: 0;
+              border: none;
+            }
             .addBtn:hover {
-                background-color: #bbb;
+              background-color: #bbb;
             }
-    </style>
-</head>
-<body>
-    <asp:ContentPlaceHolder ID="Content" runat="server">
-    </asp:ContentPlaceHolder>
-</body>
+        </style>
+    </head>
+    <body>
+        <asp:ContentPlaceHolder ID="Content" runat="server">
+
+        </asp:ContentPlaceHolder>
+    </body>
 </html>
